Resizing is working... in a way.
Mouse over the left most side of a messageArea and draw. Watch the left
side of the rect. It should move.

All surfaces will now drag, once you add them all to the update loop,
because right now, only the messageboxes are being updated per loop. along
with Surfaces I guess, but on LN149ish the size is being recalced to the
size of the image, which is problematic for dragging. Just have to work it
out after this.