reworked dragging the left side of a surface
There's still some stuff to work out, like having the surface dragged
beyond 0, so the size becomes -1 and it crashes, but I think that's
something I can fix just by adding a minimum size to the Surface class.
something like 10x10 or something. Who knows.

Anyways the issue right now is that it only drags the right side in and
out, because resizing only affects the width of the surface, and since the
left side is anchored by the bounding rect, the right side gets expanded
everytime.

I think the way to solve that is the add the diff between the new size and
old side onto the X position of the bounding box and just change that.
We'll see how it goes.