Changed the way you insert images into the application
This means that you now call a single funciton and it'll save you almost
100 characters of typing. You'll lose a bit of customization, but for now,
it doesn't matter.

Bugs found: Since MessageBox is a surface too, it gets redrawn when I call
the DrawGUISurfaces even though it's not ONLY a type. Will have to make
it so instead of just checking 'is Surface' it'll have to be none of
Surfaces children.

Will also need to use CheckMouseAgainstElements a little more elegant.
It's pretty much  spaghetti now.