Added tooltip text, which now shows up.
What I did was just add an UpdateSurface call to the loop for UpdateGUI
orchid class. I can definitely see that being unwieldy though, because I
have to make sure to have a loop for every single type of element in my
game. Right now it's tooltip, text entrys, buttons, messageboxes,
surfaces (which are just images, really)

Next, I'll just have to make the messagebox's text values so that they
don't repeat the one line of tetx over and over again. Gotta get to the
bottom of that one still.