Fixed repeating messages if count < limit
Whenever I updated a msgBox's activeMessage array, I used to have it so it
created one the size of messageLimit, which is the max number of messages
a msgBox can have displayed at once. This wouldn't work well if there was
less than that number in the total messages, because for some reason it'd
just keep putting the same number in, instead off throwing an error.

I dunno.

So just to fix it, I made it use Math.min() to find the smaller int,
either messageLimit of the length of msgList. Works out well this way.