Changed the way Surfaces are drawn, in the Orchid loop
Instead of checking if 'this is Surface', I changed to get TypeOf.Name or
something like that. That way it's children won't pass the check, and will
be drawn a second time.

Also made GuiElement.name public. I'm not sure why it wasn't before
etiher.

Next, will have to change the way I handle tooltips. because my next step
will be to just skip the draw call on the tooltips, but that'll just hide
it, rather than effectively eliminate it from the logic. That means that
when you mouse over an inactive tooltip, you'll still interact with it,
invisibly. That is no good. No good at all.