Show and hide Tooltips now. Changed 'elem is type' to 'elem.name == 'type'
I've now got it to show and hide tooltips. The issue is that it silently
throws exceptions now if there's an issue during the hoveron and hoveroff
methods. It's not an issue immediately, but it's a pretty serious design
flaw, I'm pretty sure.

Also changed the type testing in the Orchid loops to make sure a class
isn't used twice.