text is visible on menus
Made a bunch of misc changes to the code. I don't really understand how
I've set the font color up.

The reason the text was not visible on the menus was because it was
Color() rather than Color.Black. I've got to be more careful, because
COlor() is still RGBA 0 all around.

I've played around with a few random parts of the code, but I didn't make
any big changes. Mostly just deleted white space.

Sidenote, this cim console doesn't use my vimrc apparently. I dunno what's
going on with that.