Solved inheritance issue in Orchid Update/Draw calls
I'll just copy and paste my answer from SO:

It was that one of the many Parent classes I was using was marked as virtual instead of override so it incorrectly looking above the Parent class into its parent, which may have well been Object and giving me an error.

 Object.Draw
 Parent.Draw - Virtual
  Child.Draw - Override //incorrectly was virtual in my code
G_Child.Draw - Override

When I called Parent.Draw on the list, it went to Parent.Draw and saw that Child.Draw was virtual so then, I think, it went back up to Object.Draw and threw a compiler error.