Changed the Draw calls to make sure it's not hidden
Next up is adding a draw priority. I think I'll have to add a
DrawPriorityBase to all Types, and then change them around depending on
the variant of say, a menu it is. That was I can tweak which element is on
top of another of its kind. I think that'll be really good.

The harder part though will be sorting the list via draw priority though,
but maybe not. Sorting could be really easy. I'm hopeful.