Hides menus when not active inside a menu
That is to say that if you click away from a menu, it'll stop getting
drawn. Downside is it'll still get checked for mouse over. I think if I
want to stop that I'll have to add a check before I accept it as
activeElement, but before that I'd have to deal with all the menus being
in memory even though they're hidden.

Not likely to be an issue for most people, but it's still  a consideration
that should be made.

Otherwise added some comments sparingly to older code.